+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## 2.0.2
 
 - fixed default container
+- container can be injected
 
 ## 2.0.1